requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
Authlib==0.15.5
avro==1.10.0
avro-python3==1.9.2.1
azure-core==1.21.1
azure-eventhub==5.6.1
azure-eventhub-checkpointstoreblob-aio==1.1.4
azure-identity==1.7.1
azure-schemaregistry==1.0.0
azure-schemaregistry-avroserializer==1.0.0b4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
# confluent-kafka==1.7.0
cryptography==36.0.0
docopt==0.6.2
fastavro==1.4.7
Flask==2.0.2
Flask-OAuth==0.12
Flask-SQLAlchemy==2.5.1
frozenlist==1.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
greenlet==1.1.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
multidict==5.2.0
oauth2==1.9.0.post1
oauthlib==3.1.1
pipreqs==0.4.11
portalocker==1.7.1
protobuf==3.19.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.5
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.26
typing_extensions==4.0.1
uamqp==1.4.3
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
yarl==1.7.2
flask-oidc==1.4.0
okta==0.0.4
okta-jwt-verifier==0.2.0
flask-cors==3.0.9
numpy
azure-storage-queue==2.1.0